chore(build): unpack causy-ui manually as the release-downloader pipeline step has currently a bug in its extraction logic
This is a lot of fun and just costed us multiple hours. But the lib we are using to fetch the causy ui release has a bug which prevents it from extracting zips correctly.
See: https://github.com/robinraju/release-downloader/issues/689
The fix - extracting the causy ui - is super easy. But to figure out that the actual issue is just a zip extraction (and not e.g. a bug in aws agw - issue first found in causy share) was a lot of fun.

Co-authored-by: this-is-sofia <faltenbacher.sofia@gmail.com>